Fix sonar check

+++ src/creatures/npcs/npc.cpp
@@ -27,7 +27,6 @@
 #include "creatures/combat/spells.h"
 #include "lua/creature/events.h"
 
-
 int32_t Npc::despawnRange;
 int32_t Npc::despawnRadius;
 
@@ -244,7 +243,7 @@ void Npc::onPlayerBuyItem(Player* player, uint16_t itemId,
 		}
 	}
 
-	int64_t totalCost = buyPrice * amount;
+	uint32_t totalCost = buyPrice * amount;
 	if (getCurrency() == ITEM_GOLD_COIN) {
 		if (!g_game().removeMoney(player, totalCost, 0, true)) {
 			SPDLOG_ERROR("[Npc::onPlayerBuyItem (removeMoney)] - Player {} have a problem for buy item {} on shop for npc {}", player->getName(), itemId, getName());
